function-call-fastapi/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
sqlalchemy
python-dotenv
pymysql
alembic
requests